refactor layers into libraries
