+++ lib/flutter_map.dart
@@ -27,12 +27,12 @@ export 'package:flutter_map/src/layer/attribution_layer/rich/animation.dart';
 export 'package:flutter_map/src/layer/attribution_layer/rich/source.dart';
 export 'package:flutter_map/src/layer/attribution_layer/rich/widget.dart';
 export 'package:flutter_map/src/layer/attribution_layer/simple.dart';
-export 'package:flutter_map/src/layer/circle_layer.dart';
+export 'package:flutter_map/src/layer/circle_layer/circle_layer.dart';
 export 'package:flutter_map/src/layer/general/hit_detection.dart';
 export 'package:flutter_map/src/layer/general/mobile_layer_transformer.dart';
 export 'package:flutter_map/src/layer/general/translucent_pointer.dart';
-export 'package:flutter_map/src/layer/marker_layer.dart';
-export 'package:flutter_map/src/layer/overlay_image_layer.dart';
+export 'package:flutter_map/src/layer/marker_layer/marker_layer.dart';
+export 'package:flutter_map/src/layer/overlay_image_layer/overlay_image_layer.dart';
 export 'package:flutter_map/src/layer/polygon_layer/polygon_layer.dart';
 export 'package:flutter_map/src/layer/polyline_layer/polyline_layer.dart';
 export 'package:flutter_map/src/layer/tile_layer/tile_builder.dart';
