+++ lib/src/layer/circle_layer/painter.dart
@@ -1,69 +1,4 @@
-import 'dart:ui';
-
-import 'package:flutter/widgets.dart';
-import 'package:flutter_map/flutter_map.dart';
-import 'package:latlong2/latlong.dart' hide Path;
-
-/// Immutable marker options for [CircleMarker]. Circle markers are a more
-/// simple and performant way to draw markers as the regular [Marker]
-@immutable
-class CircleMarker {
-  /// An optional [Key] for the [CircleMarker].
-  /// This key is not used internally.
-  final Key? key;
-
-  /// The center coordinates of the circle
-  final LatLng point;
-
-  /// The radius of the circle
-  final double radius;
-
-  /// The color of the circle area.
-  final Color color;
-
-  /// The stroke width for the circle border. Defaults to 0 (no border)
-  final double borderStrokeWidth;
-
-  /// The color of the circle border line. Needs [borderStrokeWidth] to be > 0
-  /// to be visible.
-  final Color borderColor;
-
-  /// Set to true if the radius should use the unit meters.
-  final bool useRadiusInMeter;
-
-  /// Constructor to create a new [CircleMarker] object
-  const CircleMarker({
-    required this.point,
-    required this.radius,
-    this.key,
-    this.useRadiusInMeter = false,
-    this.color = const Color(0xFF00FF00),
-    this.borderStrokeWidth = 0.0,
-    this.borderColor = const Color(0xFFFFFF00),
-  });
-}
-
-/// A layer that displays a list of [CircleMarker] on the map
-@immutable
-class CircleLayer extends StatelessWidget {
-  /// The list of [CircleMarker]s.
-  final List<CircleMarker> circles;
-
-  /// Create a new [CircleLayer] as a child for flutter map
-  const CircleLayer({super.key, required this.circles});
-
-  @override
-  Widget build(BuildContext context) {
-    final camera = MapCamera.of(context);
-    return MobileLayerTransformer(
-      child: CustomPaint(
-        painter: CirclePainter(circles, camera),
-        size: Size(camera.size.x, camera.size.y),
-        isComplex: true,
-      ),
-    );
-  }
-}
+part of 'circle_layer.dart';
 
 /// The [CustomPainter] used to draw [CircleMarker] for the [CircleLayer].
 @immutable
