+++ test/flutter_map_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_map/src/geo/crs.dart';
 import 'package:flutter_map/src/gestures/interactive_flag.dart';
-import 'package:flutter_map/src/layer/marker_layer.dart';
+import 'package:flutter_map/src/layer/marker_layer/marker_layer.dart';
 import 'package:flutter_map/src/layer/tile_layer/tile_layer.dart';
 import 'package:flutter_map/src/map/camera/camera.dart';
 import 'package:flutter_map/src/map/controller/map_controller.dart';
